backend/requirements.txt
fastapi
uvicorn
langchain
chromadb
python-dotenv
pypdf
sentence-transformers
requests
fpdf 